+++ CHANGELOG.md
@@ -43,6 +43,8 @@ We use the following categories for changes:
   ingested using the native Jaeger API.
 - Fix traces queries returning duplicated events and links (or logs and
   references in Jaeger) for traces with more than one event and one link.
+- Fix incorrect reference types when retrieving Jaeger traces with multiple
+  parent references [#1680]. 
 
 ## [0.14.0] - 2022-08-30